Rename dom.rb to model.rb
