+++ lib/sablon/document_object_model/model.rb
